clothing-color-advisor/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
opencv-python-headless==4.10.0.84
mediapipe==0.10.33
numpy==1.26.4
scikit-learn==1.5.2
Pillow==10.4.0
